# Improve performance in `Map.Update`