--- a/Core.Tests/CommandTests.cs
+++ b/Core.Tests/CommandTests.cs
@@ -34,7 +34,7 @@ using Weikio.PluginFramework.Catalogs;
 using Xunit;
 using Xunit.Abstractions;
 
-// cannot cast MockedGameConnection to IGameConnection ??? 
+// cannot cast MockedGameConnection to IGameConnection ???
 #pragma warning disable CS8602
 
 namespace Core.Tests;
@@ -51,7 +51,6 @@ internal class MockedGameConnection : IGameConnection
     public void Close()
     {
     }
-    
 
     public Task Send<T>(T packet) where T : IPacketSerializable
     {
@@ -91,7 +90,7 @@ public class CommandTests : IAsyncLifetime
     private readonly Faker<Player> _playerDataFaker;
 
     public CommandTests(ITestOutputHelper testOutputHelper)
-    {        
+    {
         _playerDataFaker = new AutoFaker<Player>()
             .RuleFor(x => x.Level, _ => (byte)1)
             .RuleFor(x => x.St, _ => (byte)1)
@@ -122,7 +121,7 @@ public class CommandTests : IAsyncLifetime
         var dbMock = new Mock<IDbConnection>();
         dbMock.SetupDapperAsync(c => c.QueryAsync<Guid>(It.IsAny<string>(), null, null, null, null));
         _services = new ServiceCollection()
-            .AddCoreServices(new EmptyPluginCatalog())
+            .AddCoreServices(new EmptyPluginCatalog(), new ConfigurationBuilder().Build())
             .AddGameServices()
             .AddLogging(x =>
             {
@@ -177,7 +176,7 @@ public class CommandTests : IAsyncLifetime
 
         Assert.NotEmpty(_player.Inventory.Items);
         await _commandManager.Handle(_connection, "/ip");
-        
+
         Assert.Empty(_player.Inventory.Items);
     }
 
@@ -189,12 +188,12 @@ public class CommandTests : IAsyncLifetime
         await world.SpawnEntity(_player);
         await world.SpawnEntity(player2);
         await player2.Move((int)(11 * Map.MapUnit), (int)(27 * Map.MapUnit));
-        
+
         Assert.Equal((int)(10 * Map.MapUnit), _player.PositionX);
         Assert.Equal((int)(26 * Map.MapUnit), _player.PositionY);
-        
+
         await _commandManager.Handle(_connection, $"/tp \"{player2.Name}\"");
-        
+
         Assert.Equal((int)(11 * Map.MapUnit), _player.PositionX);
         Assert.Equal((int)(27 * Map.MapUnit), _player.PositionY);
     }
@@ -207,11 +206,11 @@ public class CommandTests : IAsyncLifetime
         await world.SpawnEntity(_player);
         await world.SpawnEntity(player2);
         await player2.Move((int)(11 * Map.MapUnit), (int)(27 * Map.MapUnit));
-        
-        
+
+
         Assert.Equal((int)(11 * Map.MapUnit), player2.PositionX);
         Assert.Equal((int)(27 * Map.MapUnit), player2.PositionY);
-        
+
         await _commandManager.Handle(_connection, $"/tphere \"{player2.Name}\"");
 
         Assert.Equal((int)(10 * Map.MapUnit), player2.PositionX);
@@ -247,7 +246,7 @@ public class CommandTests : IAsyncLifetime
         var player2 = ActivatorUtilities.CreateInstance<PlayerEntity>(_services, _playerDataFaker.Generate());
         await world.SpawnEntity(_player);
         await world.SpawnEntity(player2);
-        
+
         await _commandManager.Handle(_connection, $"/exp 500 \"{player2.Name}\"");
 
         player2.GetPoint(EPoints.Experience).Should().Be(500);
@@ -273,7 +272,7 @@ public class CommandTests : IAsyncLifetime
         var player2 = ActivatorUtilities.CreateInstance<PlayerEntity>(_services, _playerDataFaker.Generate());
         await world.SpawnEntity(_player);
         await world.SpawnEntity(player2);
-        
+
         await _commandManager.Handle(_connection, $"/give \"{player2.Name}\" 1 10");
 
         player2.Inventory.Items.Should().NotBeEmpty();
@@ -300,7 +299,7 @@ public class CommandTests : IAsyncLifetime
         var player2 = ActivatorUtilities.CreateInstance<PlayerEntity>(_services, _playerDataFaker.Generate());
         await world.SpawnEntity(_player);
         await world.SpawnEntity(player2);
-        
+
         player2.GetPoint(EPoints.Gold).Should().Be(0);
         await _commandManager.Handle(_connection, $"/gold 10 \"{player2.Name}\"");
         player2.GetPoint(EPoints.Gold).Should().Be(10);
@@ -316,7 +315,7 @@ public class CommandTests : IAsyncLifetime
 
         Assert.Equal((int)(10 * Map.MapUnit), _player.PositionX);
         Assert.Equal((int)(26 * Map.MapUnit), _player.PositionY);
-        
+
         await _commandManager.Handle(_connection, $"/goto {11} {27}");
 
         Assert.Equal((int)(_player.Map.PositionX + 11 * 100), _player.PositionX);
@@ -328,11 +327,11 @@ public class CommandTests : IAsyncLifetime
     {
         var world = await PrepareWorldAsync();
         await world.SpawnEntity(_player);
-        
-        
+
+
         Assert.Equal((int)(10 * Map.MapUnit), _player.PositionX);
         Assert.Equal((int)(26 * Map.MapUnit), _player.PositionY);
-        
+
         await _commandManager.Handle(_connection, "/goto --map map_b2");
 
         // target position is half of X & Y
@@ -363,7 +362,7 @@ public class CommandTests : IAsyncLifetime
         await world.SpawnEntity(player2);
 
         await _commandManager.Handle(_connection, $"/kick \"{player2.Name}\"");
-        
+
         Assert.Null(world.GetPlayer(player2.Name));
     }
 
@@ -382,7 +381,7 @@ public class CommandTests : IAsyncLifetime
     public async Task LevelCommand_Self()
     {
         _player.GetPoint(EPoints.Level).Should().Be(1);
-        
+
         await _commandManager.Handle(_connection, "/level 30");
 
         _player.GetPoint(EPoints.Level).Should().Be(30);
@@ -397,7 +396,7 @@ public class CommandTests : IAsyncLifetime
         await world.SpawnEntity(player2);
 
         player2.GetPoint(EPoints.Level).Should().Be(1);
-        
+
         await _commandManager.Handle(_connection, $"/level 30 \"{player2.Name}\"");
 
         player2.GetPoint(EPoints.Level).Should().Be(30);
@@ -408,9 +407,9 @@ public class CommandTests : IAsyncLifetime
     {
         var world = await PrepareWorldAsync();
         await world.SpawnEntity(_player);
-        
+
         world.GetPlayer(_player.Name).Should().NotBeNull();
-        
+
         await _commandManager.Handle(_connection, "/logout");
 
         world.GetPlayer(_player.Name).Should().BeNull();
@@ -421,13 +420,13 @@ public class CommandTests : IAsyncLifetime
     {
         var world = await PrepareWorldAsync();
         await world.SpawnEntity(_player);
-        
+
         world.GetPlayer(_player.Name).Should().NotBeNull();
         (_connection as MockedGameConnection).SentPhases.Should().NotContainEquivalentOf(new GCPhase
         {
             Phase = EPhases.Select
         });
-        
+
         await _commandManager.Handle(_connection, "/phase_select");
 
         _player.Connection.Phase.Should().Be(EPhases.Select);
@@ -443,9 +442,9 @@ public class CommandTests : IAsyncLifetime
     {
         var world = await PrepareWorldAsync();
         await world.SpawnEntity(_player);
-        
+
         world.GetPlayer(_player.Name).Should().NotBeNull();
-        
+
         await _commandManager.Handle(_connection, "/quit");
 
         world.GetPlayer(_player.Name).Should().BeNull();
@@ -473,7 +472,7 @@ public class CommandTests : IAsyncLifetime
         _player.Map.GetEntities().Count.Should().Be(1);
 
         await _commandManager.Handle(_connection, "/spawn 101");
-        
+
         _player.Map.GetEntities().Count.Should().Be(2);
     }
 
@@ -487,7 +486,7 @@ public class CommandTests : IAsyncLifetime
         _player.Map.GetEntities().Count.Should().Be(1);
 
         await _commandManager.Handle(_connection, "/spawn 101 10");
-        
+
         _player.Map.GetEntities().Count.Should().Be(11);
     }
 
@@ -496,9 +495,9 @@ public class CommandTests : IAsyncLifetime
     {
         await _player.AddPoint(EPoints.StatusPoints, 1);
         _player.GetPoint(EPoints.Ht).Should().Be(1);
-        
+
         await _commandManager.Handle(_connection, "/stat ht");
-        
+
         _player.GetPoint(EPoints.Ht).Should().Be(2);
     }