--- a/Core.Tests/CommandTests.cs
+++ b/Core.Tests/CommandTests.cs
@@ -8,6 +8,7 @@ using AutoBogus;
 using Bogus;
 using Dapper;
 using FluentAssertions;
+using FluentAssertions.Equivalency;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
@@ -85,6 +86,7 @@ public class CommandTests : IAsyncLifetime
     private readonly IGameConnection _connection;
     private readonly ServiceProvider _services;
     private readonly IPlayerEntity _player;
+    private readonly IItemManager _itemManager;
     private readonly Faker<Player> _playerDataFaker;
 
     public CommandTests(ITestOutputHelper testOutputHelper)
@@ -108,6 +110,8 @@ public class CommandTests : IAsyncLifetime
         itemManagerMock.Setup(x => x.GetItem(It.IsAny<uint>())).Returns<uint>(id => new AutoFaker<ItemData>()
             .RuleFor(x => x.Id, _ => id)
             .RuleFor(x => x.Size, _ => (byte)1)
+            .RuleFor(x => x.WearFlags, _ => (byte)EWearFlags.Weapon)
+            .RuleFor(x => x.Values, _ => new List<int>{0, 0, 0, 10, 16, 0})
             .Generate());
         var cacheManagerMock = new Mock<ICacheManager>();
         var redisListWrapperMock = new Mock<IRedisListWrapper<Guid>>();
@@ -145,6 +149,7 @@ public class CommandTests : IAsyncLifetime
             .AddSingleton<IPlayerEntity, PlayerEntity>()
             .AddSingleton(_ => _playerDataFaker.Generate())
             .BuildServiceProvider();
+        _itemManager = _services.GetRequiredService<IItemManager>();
         _commandManager = _services.GetRequiredService<ICommandManager>();
         _commandManager.Register("QuantumCore.Game.Commands", typeof(SpawnCommand).Assembly);
         _connection = _services.GetRequiredService<IGameConnection>();
@@ -185,6 +190,7 @@ public class CommandTests : IAsyncLifetime
         var player2 = ActivatorUtilities.CreateInstance<PlayerEntity>(_services, _playerDataFaker.Generate());
         world.SpawnEntity(_player);
         world.SpawnEntity(player2);
+        world.Update(0); // spawn entities
         player2.Move((int)(11 * Map.MapUnit), (int)(27 * Map.MapUnit));
 
         Assert.Equal((int)(10 * Map.MapUnit), _player.PositionX);
@@ -203,6 +209,7 @@ public class CommandTests : IAsyncLifetime
         var player2 = ActivatorUtilities.CreateInstance<PlayerEntity>(_services, _playerDataFaker.Generate());
         world.SpawnEntity(_player);
         world.SpawnEntity(player2);
+        world.Update(0); // spawn entities
         player2.Move((int)(11 * Map.MapUnit), (int)(27 * Map.MapUnit));
 
 
@@ -218,15 +225,21 @@ public class CommandTests : IAsyncLifetime
     [Fact]
     public async Task DebugCommand()
     {
-        await _commandManager.Handle(_connection, "debug_damage");
+        var item = new ItemInstance{ItemId = 1, Count = 1};
+        var wearSlot = _player.Inventory.EquipmentWindow.GetWearPosition(_itemManager, item);
 
+        _player.SetItem(item, (byte) WindowType.Inventory, (ushort)wearSlot);
+
+        await _commandManager.Handle(_connection, "debug_damage");
         // simple calculation just for this test
-        var minAttack = _player.GetPoint(EPoints.Level) + _player.GetPoint(EPoints.St);
-        var maxAttack = _player.GetPoint(EPoints.Level) + _player.GetPoint(EPoints.St);
-        const int minWeapon = 0;
-        const int maxWeapon = 0;
-        (_connection as MockedGameConnection).SentMessages.Should().ContainEquivalentOf(new ChatOutcoming { Message = $"Weapon Damage: {minWeapon}-{maxWeapon}" }, cfg => cfg.Including(x => x.Message));
-        (_connection as MockedGameConnection).SentMessages.Should().ContainEquivalentOf(new ChatOutcoming { Message = $"Attack Damage: {minAttack}-{maxAttack}" }, cfg => cfg.Including(x => x.Message));
+        var minAttack = _player.GetPoint(EPoints.MinAttackDamage);
+        var maxAttack = _player.GetPoint(EPoints.MaxAttackDamage);
+        var sentMessages = (_connection as MockedGameConnection).SentMessages;
+
+        sentMessages.Should().ContainEquivalentOf(new ChatOutcoming { Message = $"Weapon Damage: 10-16" }, Config);
+        sentMessages.Should().ContainEquivalentOf(new ChatOutcoming { Message = $"Attack Damage: {minAttack}-{maxAttack}" }, Config);
+
+        EquivalencyAssertionOptions<ChatOutcoming> Config(EquivalencyAssertionOptions<ChatOutcoming> cfg) => cfg.Including(x => x.Message);
     }
 
     [Fact]
@@ -308,6 +321,7 @@ public class CommandTests : IAsyncLifetime
     {
         var world = await PrepareWorldAsync();
         world.SpawnEntity(_player);
+        world.Update(0); // spawn entities
 
         _player.Move((int)(Map.MapUnit * 10), (int)(Map.MapUnit * 26));
 
@@ -325,6 +339,7 @@ public class CommandTests : IAsyncLifetime
     {
         var world = await PrepareWorldAsync();
         world.SpawnEntity(_player);
+        world.Update(0); // spawn entities
 
 
         Assert.Equal((int)(10 * Map.MapUnit), _player.PositionX);
@@ -465,11 +480,13 @@ public class CommandTests : IAsyncLifetime
     {
         var world = await PrepareWorldAsync();
         world.SpawnEntity(_player);
+        world.Update(0); // spawn entities
         _player.Move((int)(Map.MapUnit * 13), (int)(Map.MapUnit * 29)); // center of the map
         await File.WriteAllTextAsync("settings.toml", @"maps = [""map_a2"", ""map_b2""]");
         _player.Map.Entities.Count.Should().Be(1);
 
         await _commandManager.Handle(_connection, "/spawn 101");
+        world.Update(0); // spawn entities
 
         _player.Map.Entities.Count.Should().Be(2);
     }
@@ -479,11 +496,13 @@ public class CommandTests : IAsyncLifetime
     {
         var world = await PrepareWorldAsync();
         world.SpawnEntity(_player);
+        world.Update(0); // spawn entities
         _player.Move((int)(Map.MapUnit * 13), (int)(Map.MapUnit * 29)); // center of the map
         await File.WriteAllTextAsync("settings.toml", @"maps = [""map_a2"", ""map_b2""]");
         _player.Map.Entities.Count.Should().Be(1);
 
         await _commandManager.Handle(_connection, "/spawn 101 10");
+        world.Update(0); // spawn entities
 
         _player.Map.Entities.Count.Should().Be(11);
     }