--- a/Core/Extensions/ConnectionExtensions.cs
+++ b/Core/Extensions/ConnectionExtensions.cs
@@ -7,10 +7,10 @@ namespace QuantumCore.Extensions;
 
 public static class ConnectionExtensions
 {
-    public static async Task SetPhaseAsync(this IConnection connection, EPhases phase)
+    public static void SetPhaseAsync(this IConnection connection, EPhases phase)
     {
         connection.Phase = phase;
-        await connection.Send(new GCPhase
+        connection.Send(new GCPhase
         {
             Phase = connection.Phase
         });