--- a/Game.Tests/Usings.cs
+++ b/Game.Tests/Usings.cs
@@ -1,0 +1,1 @@
+global using Xunit;