… Core.Tests/ParserTests.cs
… Core/Core/Networking/ServerBase.cs
… Core/Extensions/ServiceExtensions.cs
… Core/QuantumCoreHostBuilder.cs
… CorePluginAPI/Core/Models/PlayerData.cs
… CorePluginAPI/Game/World/AI/IBehaviour.cs
… CorePluginAPI/Game/World/IEntity.cs
… CorePluginAPI/Game/World/IMap.cs
… CorePluginAPI/Game/World/IPlayerEntity.cs
… CorePluginAPI/Game/World/IWorld.cs
… CorePluginAPI/Game/World/SpawnGroup.cs
… CorePluginAPI/IConnection.cs
… CorePluginAPI/IEquipment.cs
… CorePluginAPI/IGroundItem.cs
… CorePluginAPI/IInventory.cs
… CorePluginAPI/SlotChangedEventArgs.cs
… Executables/Auth/AuthServer.cs
… Executables/Auth/PacketHandlers/LoginRequestHandler.cs
… Executables/Game/ChatManager.cs
… Executables/Game/Commands/ClearInventoryCommand.cs
… Executables/Game/Commands/CommandCache.cs
… Executables/Game/Commands/CommandManager.cs
… Executables/Game/Commands/CommandTeleportHere.cs
… Executables/Game/Commands/CommandTeleportTo.cs
… Executables/Game/Commands/DebugCommand.cs
… Executables/Game/Commands/ExperienceCommand.cs
… Executables/Game/Commands/GiveItemCommand.cs
… Executables/Game/Commands/GoldCommand.cs
… Executables/Game/Commands/GotoCommand.cs
… Executables/Game/Commands/HpCommand.cs