@rausc-daniel 2023-10-22
Very impressive improvements indeed 💪
If you manage to clean up the calls to Task.Wait() and Task.Result that for now seem to not be easily avoidable it would be even more impressive ;)